Measures cannot work alone, folders are necessary
